Add feature-flagged ability to show chance % on covered mines

+++ app/src/main/java/com/lordinatec/minesweepercompose/minesweeper/apis/model/FieldViewState.kt
@@ -7,7 +7,7 @@ data class FieldViewState(
     val gameOver: Boolean = false,
     val winner: Boolean = false,
     val minesRemaining: Int = Config.MINES,
-    val timeValue: Long = 0L,
+    val timeValue: Float = 0f,
     val tileStates: List<TileState> = List<TileState>(Config.WIDTH * Config.HEIGHT) { TileState.COVERED },
     val tileValues: List<String> = List<String>(Config.WIDTH * Config.HEIGHT) { "" }
 )
